requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.8
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
conda-pack==0.7.1
cycler==0.12.1
datasets==1.18.3
decorator==5.1.1
dill==0.3.8
einops==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.1.0
gpustat==1.1
higher==0.2.1
huggingface-hub==0.21.4
hydra-core==1.1.1
idna==3.6
importlib-metadata==6.3.0
iopath==0.1.10
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
kiwisolver==1.4.5
lightning-utilities==0.10.1
lit==17.0.6
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.6.5
numpy==1.22.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==2.1.1
openai==0.27.9
opencv-python==4.8.0.76
packaging==23.2
pandas==1.4.0
parso==0.8.3
peft==0.7.1
pexpect==4.9.0
pillow==10.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2024.1
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
timm==0.9.7
tokenizers==0.13.3
torchmetrics==0.9.3
tqdm==4.62.3
traitlets==5.14.1
transformers==4.30.1
triton==2.0.0
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0